refactor: move registerUser to logic; update websocket_server/index.ts
